backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
alembic
pydantic
python-jose[cryptography]
python-dotenv
pydantic-settings
celery
redis
paho-mqtt